Add .NET Aspire for application orchestration (#18)
* add unversioned schemas file

* futz about with ide config files

* fix c# style of sln capitalization

* initial aspireification

* Decouple build from DB connections

* add missing ide config items

* Add optional 'SEED' flag for db seeding

* Update README to cover aspire and seed

* fix idea .gitignore

* clean up ide configs

* Fix service discovery and force port so auth0 callback works

* fix api envvar

* update prerequisites

* fix postgres data persistence

* add documentation for PowerShell support

+++ LineUp.Backend/Program.cs
@@ -49,13 +49,7 @@ builder.Services.AddControllers();
 builder.Services.AddOpenApi();
 builder.Services.AddSingleton<IAuthorizationHandler, HasScopeHandler>();
 
-builder.Services.AddDbContext<LineUpContext>(options =>
-{
-    options.UseNpgsql(
-        builder.Configuration.GetConnectionString("LineUp")
-            ?? "Host=localhost;Username=postgres;Password=postgres;Database=my-database"
-    );
-});
+builder.AddNpgsqlDbContext<LineUpContext>("postgresdb");
 
 var app = builder.Build();
 
@@ -90,8 +84,11 @@ app.UseAuthentication();
 app.UseAuthorization();
 app.MapControllers();
 
-using (var scope = app.Services.CreateScope())
+var seed = app.Environment.IsDevelopment() && Environment.GetEnvironmentVariable("SEED") == "true";
+
+if (seed)
 {
+    using var scope = app.Services.CreateScope();
     var db = scope.ServiceProvider.GetRequiredService<LineUpContext>();
     //TODO DONT DO THIS IN PROD!!!!!!!!!!!!!! :(((((
     db.Database.EnsureDeleted();
